Add LoadKey tracking and unload on ~ctor
Leaving a loaded hive causes the ownership of the loaded subkey to
transfer to the system process.  You can't unload it again without a
restart.

RegLoadAppKey seem the better solution, however I have troubles with it
reporting a corrupted database when regedit loads it just fine.  Almost
no documentation online.

Added the GetRegValue back in for systems that can use it.  It is a
better solution, just not available for the older systems.